requirements.txt
# requirements.txt
# Sentiment-Augmented VIX Forecasting
# Python 3.10+

# Core Data Science
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# NLP & Sentiment Analysis
transformers==4.36.2
torch==2.1.2
sentencepiece==0.1.99

# Time Series & Econometrics
statsmodels==0.14.1

# Machine Learning
scikit-learn==1.3.2

# Data Collection
yfinance==0.2.33
requests==2.31.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
